Use @graphql-ts/schema
