+++ packages-next/fields/src/types/autoIncrement/index.ts
@@ -67,7 +67,6 @@ export const autoIncrement =
         },
         output: types.field({
           type: types.nonNull(types.ID),
-          // TODO: should @ts-gql/schema understand the coercion that graphql-js can do here?
           resolve({ value }) {
             return value.toString();
           },
