+++ packages-next/types/package.json
@@ -8,8 +8,8 @@
     "node": "^12.20 || >= 14.13"
   },
   "dependencies": {
+    "@graphql-ts/schema": "0.1.1",
     "@keystone-next/fields": "^10.0.0",
-    "@ts-gql/schema": "0.11.2",
     "apollo-server-types": "^0.9.0",
     "cors": "^2.8.5",
     "decimal.js": "^10.2.1",
