+++ packages-next/types/src/schema/index.ts
@@ -1,0 +1,1 @@
+export * as types from './graphql-ts-schema';
