+++ packages-next/types/src/schema/schema-api-with-context.d.ts
@@ -1,0 +1,6 @@
+import { SchemaAPIWithContext } from '@graphql-ts/schema';
+import { Context } from './graphql-ts-schema';
+
+declare const __schema: SchemaAPIWithContext<Context>;
+
+export = __schema;
