+++ packages-next/types/src/schema/schema-api-with-context.js
@@ -1,0 +1,1 @@
+export * from '@graphql-ts/schema/api-with-context';
