--- packages-next/types/src/ts-gql-schema/index.ts
@@ -1,1 +1,0 @@
-export * as types from './ts-gql-schema';
