--- packages-next/types/src/ts-gql-schema/types-with-context.d.ts
@@ -1,6 +1,0 @@
-import { TypesWithContext } from '@ts-gql/schema';
-import { Context } from './ts-gql-schema';
-
-declare const __types: TypesWithContext<Context>;
-
-export = __types;
