--- packages-next/types/src/ts-gql-schema/types-with-context.js
@@ -1,1 +1,0 @@
-export * from '@ts-gql/schema/types-with-context';
